Extract `collection/shard_transfer.rs`
